Fix: rotas de fazenda toleram colunas de billing ausentes no banco de produção
O banco real está sem as colunas Farm.stripeSubscriptionItemId e
Farm.stripePlanTier que o schema.prisma já declara, então qualquer
include/select geral de Farm quebrava com P2022 em praticamente toda rota do
dashboard (login, listagem de animais, edição de animal, convites de equipe,
checkout e webhook do Stripe). Passa a selecionar explicitamente só as colunas
existentes no contexto de fazenda (tenant.ts, farms/ensure) e isola a leitura/
escrita dos campos de billing por assento atrás de helpers com fallback
seguro (stripeSeats.ts), usados em farmSessions e farm-invites.

Também corrige a truncagem do peso ao editar o animal (parseInt -> parseFloat)
e remove uma segunda busca redundante de vacinas na tela de detalhe do animal,
que já vinham incluídas na consulta principal.

+++ src/app/dashboard/[id]/(components)/editableAnimalDetails.tsx
@@ -709,7 +709,7 @@ const EditableAnimalDetails: React.FC<EditableAnimalDetailsProps> = ({
       type === 'checkbox'
         ? (event.target as HTMLInputElement).checked
         : type === 'number' || type === 'range'
-          ? parseInt(value)
+          ? parseFloat(value)
           : value;
     setAllDataForm((prev) => ({ ...prev, [name]: newValue }));
   };
